# docs/mobile_v1_openapi.yaml
openapi: 3.0.3
info:
  title: FichaYa - Mobile API v1
  version: 1.25.0
  description: |
    API REST para la aplicacion Flutter de control de asistencia.
    Todos los endpoints (excepto auth) requieren header `Authorization: Bearer <token>`.
    Los modulos complementarios Feedback (`/api/v1/feedback`) y SKAP (`/api/skap`)
    usan el mismo Bearer JWT mobile, aunque no cuelgan del prefijo `/api/v1/mobile`.
    La campana de Alertas del home agrupa adelantos, pedidos de mercaderia y feedback
    para mostrar novedades o aprobaciones pendientes sin cambiar los flujos existentes.
  contact:
    name: Equipo Backend
servers:
  - url: http://localhost:5000/api/v1/mobile
    description: Local
  - url: https://control-asistencia-backend-8gle.onrender.com/api/v1/mobile
    description: Produccion

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JustificacionItem:
      type: object
      properties:
        id:
          type: integer
        asistencia_id:
          type: integer
          nullable: true
        asistencia_fecha:
          type: string
          format: date
          nullable: true
        fecha_desde:
          type: string
          format: date
        fecha_hasta:
          type: string
          format: date
        motivo:
          type: string
        archivo:
          type: string
          nullable: true
        legajo_evento_id:
          type: integer
          nullable: true
        adjuntos_count:
          type: integer
          default: 0
        adjuntos_max:
          type: integer
          example: 10
        adjuntos_disponibles:
          type: integer
        adjuntos:
          type: array
          items:
            $ref: '#/components/schemas/JustificacionAdjuntoItem'
        estado:
          type: string
          enum: [pendiente, aprobada, rechazada]
          default: pendiente
        resuelto_at:
          type: string
          format: date-time
          nullable: true
        resuelto_by_usuario_id:
          type: integer
          nullable: true
        resuelto_by_usuario:
          type: string
          nullable: true
        comentario_resolucion:
          type: string
          nullable: true
        motivo_rechazo:
          type: string
          nullable: true
        notificado_empleado_at:
          type: string
          format: date-time
          nullable: true
        visto_por_empleado_at:
          type: string
          format: date-time
          nullable: true
        tiene_novedad:
          type: boolean
        created_at:
          type: string
          format: date-time
    JustificacionAdjuntoItem:
      type: object
      properties:
        id:
          type: integer
        evento_id:
          type: integer
        nombre_original:
          type: string
        mime_type:
          type: string
        extension:
          type: string
          nullable: true
        tamano_bytes:
          type: integer
        estado:
          type: string
        created_at:
          type: string
          format: date-time
        download_url:
          type: string
    JustificacionCreateRequest:
      type: object
      required: [fecha_desde, fecha_hasta, motivo]
      properties:
        fecha_desde:
          type: string
          format: date
        fecha_hasta:
          type: string
          format: date
        asistencia_id:
          type: integer
          nullable: true
        motivo:
          type: string
          example: "Enfermedad con certificado medico"
        archivo:
          type: string
          nullable: true
          example: "https://.../cert.pdf"
    JustificacionCreateMultipartRequest:
      type: object
      required: [fecha_desde, fecha_hasta, motivo]
      properties:
        fecha_desde:
          type: string
          format: date
        fecha_hasta:
          type: string
          format: date
        asistencia_id:
          type: integer
          nullable: true
        motivo:
          type: string
        archivo:
          type: string
          nullable: true
        adjuntos:
          type: array
          maxItems: 10
          items:
            type: string
            format: binary
    JustificacionUpdateRequest:
      type: object
      required: [motivo]
      properties:
        motivo:
          type: string
        archivo:
          type: string
          nullable: true
    JustificacionUpdateMultipartRequest:
      type: object
      required: [motivo]
      properties:
        motivo:
          type: string
        archivo:
          type: string
          nullable: true
        adjuntos:
          type: array
          items:
            type: string
            format: binary
    PaginatedJustificaciones:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/JustificacionItem'
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string

    # --- Feedback de calle ---
...
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /me/justificaciones:
    get:
      tags: [Justificaciones]
      summary: Lista paginada de justificaciones del empleado
      parameters:
        - name: desde
          in: query
          schema:
            type: string
            format: date
        - name: hasta
          in: query
          schema:
            type: string
            format: date
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendiente, aprobada, rechazada]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Lista de justificaciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedJustificaciones'
    post:
      tags: [Justificaciones]
      summary: Crea una justificacion (estado inicial siempre pendiente)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JustificacionCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/JustificacionCreateMultipartRequest'
      responses:
        '201':
          description: Justificacion creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JustificacionItem'
        '400':
          description: Validacion fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/justificaciones/{id}:
    get:
      tags: [Justificaciones]
      summary: Detalle de una justificacion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Justificacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JustificacionItem'
        '404':
          description: No encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Justificaciones]
      summary: Edita una justificacion (solo si esta pendiente)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JustificacionUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/JustificacionUpdateMultipartRequest'
      responses:
        '200':
          description: Justificacion actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JustificacionItem'
        '404':
          description: No encontrada
        '409':
          description: No esta en estado pendiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Justificaciones]
      summary: Retira una justificacion (solo si esta pendiente)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Eliminada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '404':
          description: No encontrada
        '409':
          description: No esta en estado pendiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...